chore: useOrganizationsQuery, useDepartmentsQuery 훅 위치 변경
