+++ src/hooks/tanstack-query/committee/sign-up/index.ts
@@ -1,38 +1,8 @@
-import { MutateOptions, useMutation, useQuery } from "@tanstack/react-query";
-import { getDepartments, getOrganizations, postCommitteeSignUp } from "@/apis/committee/sign-up";
+import { MutateOptions, useMutation } from "@tanstack/react-query";
+import { postCommitteeSignUp } from "@/apis/committee/sign-up";
 import { CommitteeSignUpFormData } from "@/types/committee/sign-up";
 import { useRouter } from "next/navigation";
 import { ROUTE } from "@/constants/routes";
-
-export const useOrganizationsQuery = (type: "학생회" | "위원회" | "값을 선택해주세요.") => {
-  let category: "COUNCIL" | "COMMITTEE";
-
-  if (type === "학생회") {
-    category = "COUNCIL";
-  } else if (type === "위원회") {
-    category = "COMMITTEE";
-  }
-
-  const { data: organizations } = useQuery({
-    queryKey: ["organizations", type],
-    queryFn: () => getOrganizations(category),
-    staleTime: 1000 * 60 * 60,
-    enabled: type === "학생회" || type === "위원회",
-  });
-
-  return organizations;
-};
-
-export const useDepartmentsQuery = (id: number) => {
-  const { data: departments } = useQuery({
-    queryKey: ["departments", id],
-    queryFn: () => getDepartments(id),
-    staleTime: 1000 * 60 * 60,
-    enabled: !!id,
-  });
-
-  return departments;
-};
 
 export const useCommitteeSignUp = () => {
   const router = useRouter();
